+++ packages/windie-sdk-js/src/index.ts
@@ -12,6 +12,7 @@ export * from './runtime/WindieAgent.js';
 export * from './runtime/WindieClient.js';
 export * from './runtime/LocalSidecarRuntime.js';
 export * from './transport/backendEventNormalizer.js';
+export * from './transport/BackendSocketFactory.js';
 export * from './transport/HostedBackendHttpClient.js';
 export * from './transport/ManagedBackendSession.js';
 export * from './tools/ToolExecutionCoordinator.js';